minor project/medi_vault/requirements.txt
# MediVault — Flask + Firebase (Python 3.11+)
Flask>=2.3,<4
firebase-admin>=6.5
google-cloud-storage>=2.14
web3>=6.11
cryptography>=42.0
sendgrid>=6.11
python-dotenv>=1.0
requests>=2.31

# Production
gunicorn==21.2.0
flask-talisman==1.1.0
flask-limiter==3.5.0
redis==5.0.1

# Dev / test
pytest>=8.0
pytest-mock>=3.14
